backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
sqlalchemy>=2.0.27
python-dotenv>=1.0.1
requests>=2.31.0
pydantic>=2.6.3
pandas